requirements.txt
python-telegram-bot>=20.0
pytz>=2023.3
APScheduler>=3.10.0
openpyxl>=3.1.0
